requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-whois==0.8.0
shodan==1.30.1
censys==2.2.8
openai==1.3.7
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
click==8.1.7
pyfiglet==1.0.2
